Updating RateTimeInTransit based on PR#145. Also moved EstimatedArrival and ServiceSummary logic to traits because RateTimeInTransit uses slightly a different xml schema (for BusienssDaysInTransit vs BusinessTransitDays)
